DRY up post mixin/optional session
