+++ modules/exploits/windows/smb/psexec.rb
@@ -69,6 +69,7 @@ class MetasploitModule < Msf::Exploit::Remote
           [ 'Command', { 'Arch' => [ARCH_CMD], 'Payload' => { 'Space' => 8191 } } ]
         ],
       'DefaultTarget'  => 0,
+      'SessionTypes' => %w[SMB],
       # For the CVE, PsExec was first released around February or March 2001
       'DisclosureDate' => '1999-01-01'
     ))
